Refactor: Split route integration suite by domain
